refactor: Rename provides_page.ts to providers_page.ts
